+++ .eslintrc.js
@@ -1,49 +1,27 @@
-/* @flow */
-
 module.exports = {
-  'parser': '@babel/eslint-parser',
-  'env': {
-    'browser': true,
-    'jest': true,
-    'node': true,
-    'es6': true
+  root: true,
+  env: {
+    browser: true,
+    jest: true,
+    node: true,
+    es6: true,
   },
-  'extends': ['eslint:recommended', 'plugin:react/recommended'],
-  'parserOptions': {
-    'ecmaFeatures': {
-      'experimentalObjectRestSpread': true,
-      'jsx': true
-    },
-    'sourceType': 'module'
-  },
-  'plugins': [
-    'react', 'flowtype'
+  extends: [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:react/recommended",
+    "plugin:react/jsx-runtime",
   ],
-  'settings': {
-    'react': {
-      'version': '16.6',
-      'flowVersion': '0.85'
-    }
+  settings: {
+    react: {
+      version: "18.0",
+    },
   },
-  'rules': {
-    'flowtype/define-flow-type': 1,
-    'flowtype/require-valid-file-annotation': ['error', 'always'],
-
-    'react/no-unescaped-entities': ['off'],
-
-    'no-unused-vars': ['error', { 'args': 'none' }],
-
-    'indent': ['error', 2],
-    'linebreak-style': ['error', 'unix'],
-    'quotes': ['error', 'single', 'avoid-escape'],
-    'semi': ['error', 'always'],
-    'no-var': ['error'],
-    'brace-style': ['error'],
-    'array-bracket-spacing': ['error', 'never'],
-    'block-spacing': ['error', 'always'],
-    'no-spaced-func': ['error'],
-    'no-whitespace-before-property': ['error'],
-    'space-before-blocks': ['error', 'always'],
-    'keyword-spacing': ['error']
-  }
+  rules: {
+    "react/no-unescaped-entities": ["off"],
+    "@typescript-eslint/no-explicit-any": ["off"],
+    "@typescript-eslint/no-non-null-assertion": ["off"],
+    "@typescript-eslint/no-unused-vars": ["error", { args: "none" }],
+    "no-var": ["error"],
+  },
 };